requirements.txt
# Signal Hub — Python dependencies

# RSS/Atom parsing
feedparser>=6.0.11

# HTTP requests
requests>=2.31.0

# AI synthesis
anthropic>=0.39.0

# Configuration
pyyaml>=6.0.1

# Date/time handling
python-dateutil>=2.8.2

# HTML parsing (for content extraction)
beautifulsoup4>=4.12.2
lxml>=5.1.0

# Web search
duckduckgo-search>=7.0.0

# Jinja2 (used indirectly; kept for compatibility)
jinja2>=3.1.4
